+++ src/components/Link/__tests__/Link.spec.js
@@ -1,0 +1,60 @@
+import '@testing-library/jest-dom';
+import { render } from '@testing-library/vue';
+import { createRouter, createMemoryHistory } from 'vue-router';
+import { constants } from '@/config';
+import LobLink from '../Link.vue';
+
+const initialProps = {
+  to: '/settings'
+};
+
+const routes = [
+  { path: '', component: { template: '<div></div>' } },
+  { path: '/', component: { template: '<div>/</div>' } },
+  { path: '/settings', component: { template: '<div>settings</div>' } }
+];
+const router = createRouter({
+  history: createMemoryHistory(),
+  routes
+});
+
+const renderComponent = async (options) => {
+  const result = render(LobLink, {
+    ...options,
+    global: { plugins: [router, constants] }
+  });
+  await router.isReady();
+  return result;
+};
+
+describe('Link', () => {
+
+  it('renders correctly when an internal link', async () => {
+    const props = initialProps;
+    const { queryByRole } = await renderComponent({ props });
+
+    const link = queryByRole('link');
+    expect(link).toBeInTheDocument();
+  });
+
+  it('renders correctly when an external link', async () => {
+    const props = {
+      ...initialProps,
+      to: 'https://google.com'
+    };
+    const { queryByRole } = await renderComponent({ props });
+
+    const link = queryByRole('link');
+    expect(link).toBeInTheDocument();
+  });
+
+  it('renders slot content correctly', async () => {
+    const slotContent = 'Hello, this is a link.';
+    const slots = { default: slotContent };
+    const { queryByText } = await renderComponent({ slots });
+
+    const linkContent = queryByText(slotContent);
+    expect(linkContent).toBeInTheDocument();
+  });
+
+});
